requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
scikit-image>=0.21.0
pytest>=8.0.0
gdown>=5.1.0
python-dotenv>=1.0.0
